backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy==2.0.36
pymysql==1.1.1
cryptography==44.0.0
pydantic==2.10.6
pydantic-settings==2.7.1
python-dotenv==1.0.1
apscheduler==3.10.4
httpx==0.28.1
python-multipart==0.0.20
aiofiles==24.1.0
psutil==5.9.8
aiohttp==3.13.3